+++ Fyber/src/main/java/com/applovin/mediation/adapters/InneractiveMediationAdapter.java
@@ -43,6 +43,8 @@ import com.fyber.inneractive.sdk.external.VideoContentListener;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import androidx.annotation.Nullable;
+
 import static com.applovin.sdk.AppLovinSdkUtils.isValidString;
 
 public class InneractiveMediationAdapter
@@ -63,7 +65,7 @@ public class InneractiveMediationAdapter
     public InneractiveMediationAdapter(final AppLovinSdk sdk) { super( sdk ); }
 
     @Override
-    public void initialize(final MaxAdapterInitializationParameters parameters, final Activity activity, final OnCompletionListener onCompletionListener)
+    public void initialize(final MaxAdapterInitializationParameters parameters, @Nullable final Activity activity, final OnCompletionListener onCompletionListener)
     {
         if ( initialized.compareAndSet( false, true ) )
         {
@@ -143,7 +145,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void collectSignal(final MaxAdapterSignalCollectionParameters parameters, final Activity activity, final MaxSignalCollectionListener callback)
+    public void collectSignal(final MaxAdapterSignalCollectionParameters parameters, @Nullable final Activity activity, final MaxSignalCollectionListener callback)
     {
         log( "Collecting signal..." );
 
@@ -154,7 +156,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void loadInterstitialAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxInterstitialAdapterListener listener)
+    public void loadInterstitialAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxInterstitialAdapterListener listener)
     {
         log( "Loading " + ( isValidString( parameters.getBidResponse() ) ? "bidding " : "" ) + "interstitial ad for spot id \"" + parameters.getThirdPartyAdPlacementId() + "\"..." );
 
@@ -251,7 +253,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void showInterstitialAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxInterstitialAdapterListener listener)
+    public void showInterstitialAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxInterstitialAdapterListener listener)
     {
         log( "Showing interstitial ad..." );
 
@@ -268,7 +270,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void loadRewardedAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxRewardedAdapterListener listener)
+    public void loadRewardedAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxRewardedAdapterListener listener)
     {
         log( "Loading " + ( isValidString( parameters.getBidResponse() ) ? "bidding " : "" ) + "rewarded ad for spot id \"" + parameters.getThirdPartyAdPlacementId() + "\"..." );
 
@@ -406,7 +408,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void showRewardedAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxRewardedAdapterListener listener)
+    public void showRewardedAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxRewardedAdapterListener listener)
     {
         log( "Showing rewarded ad..." );
 
@@ -426,7 +428,7 @@ public class InneractiveMediationAdapter
     }
 
     @Override
-    public void loadAdViewAd(final MaxAdapterResponseParameters parameters, final MaxAdFormat adFormat, final Activity activity, final MaxAdViewAdapterListener listener)
+    public void loadAdViewAd(final MaxAdapterResponseParameters parameters, final MaxAdFormat adFormat, @Nullable final Activity activity, final MaxAdViewAdapterListener listener)
     {
         log( "Loading " + ( isValidString( parameters.getBidResponse() ) ? "bidding " : "" ) + adFormat.getLabel() + " ad for spot id \"" + parameters.getThirdPartyAdPlacementId() + "\"..." );
 
@@ -506,11 +508,20 @@ public class InneractiveMediationAdapter
             @Override
             public void onInneractiveSuccessfulAdRequest(final InneractiveAdSpot inneractiveAdSpot)
             {
+                // Extract a local variable to avoid a race condition where adViewGroup could be destroyed while this callback is invoked
+                final ViewGroup currentAdViewGroup = adViewGroup;
+                if ( currentAdViewGroup == null )
+                {
+                    log( "AdView container destroyed before it could be loaded" );
+                    listener.onAdViewAdLoadFailed( MaxAdapterError.INVALID_LOAD_STATE );
+                    return;
+                }
+
                 if ( inneractiveAdSpot.isReady() )
                 {
                     log( "AdView loaded" );
-                    controller.bindView( adViewGroup );
-                    listener.onAdViewAdLoaded( adViewGroup );
+                    controller.bindView( currentAdViewGroup );
+                    listener.onAdViewAdLoaded( currentAdViewGroup );
                 }
                 else
                 {
@@ -643,7 +654,7 @@ public class InneractiveMediationAdapter
         return new MaxAdapterError( adapterError.getErrorCode(), adapterError.getErrorMessage(), adapterErrorCode, adapterErrorStr );
     }
 
-    private Context getContext(Activity activity)
+    private Context getContext(@Nullable final Activity activity)
     {
         // NOTE: `activity` can only be null in 11.1.0+, and `getApplicationContext()` is introduced in 11.1.0
         return ( activity != null ) ? activity.getApplicationContext() : getApplicationContext();